Enemy damage
The enemy can now receive damage if hit with a trigger that has a PlayerMelee or PlayerRanged tag and the Damage script.  Also set up the enemy so it activates a collider that can deal damage to the player in a similar manner once the player is set up to take damage.  A "GhostEnemy" prefab has been created as well.