Camera, Player Animation Timing, Player Damage
I've modified the camera such that the camera only moves if the mouse is
outside the center area of the screen. I've also adjusted player
animation timing somewhat for fluidity. I've shamelessly (but not
without gratitude) integrated Kaitlyn's damage system into the player
controller with the slight change of making the player animations an
array due to multiple attack animations for the player.